Added reset button and cleaned up js file
Created a reset button on top of page which allows user to enter a new
grid size value using ints from 0 to 100. Any other values will produce
an error text on the bottom of the page and require the user to enter
another value. Removed old comments from the js file which were no
longer needed.